+++ packages/web/components/elements/Recaptcha.tsx
@@ -1,0 +1,19 @@
+import { GoogleReCaptchaCheckbox } from '@google-recaptcha/react'
+
+type RecaptchaProps = {
+  setRecaptchaToken: (token: string) => void
+}
+
+export const Recaptcha = (props: RecaptchaProps): JSX.Element => {
+  return (
+    <>
+      <GoogleReCaptchaCheckbox
+        key="recaptcha"
+        onChange={(token) => {
+          console.log('recaptcha: ', token)
+          props.setRecaptchaToken(token)
+        }}
+      />
+    </>
+  )
+}
